home-schooling/backend/requirements.txt
aniso8601==9.0.1
astroid==3.0.2
asttokens==2.4.1
black==23.11.0
blinker==1.7.0
boto3==1.34.19
botocore==1.34.19
cachelib==0.9.0
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
dill==0.3.7
et-xmlfile==1.1.0
executing==2.0.1
Flask==3.0.0
Flask-Caching==2.3.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Flask-Mail-SendGrid==0.6
Flask-RESTful==0.3.10
fpdf==1.7.2
idna==3.6
ipython==8.24.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
loguru==0.7.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector-python==8.4.0
nltk==3.8.1
numpy==1.26.4
packaging==23.2
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==4.21.12
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==2.8.0
pylint==3.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2024.1
redis==5.0.4
regex==2024.5.15
reportlab==4.0.9
requests==2.31.0
s3transfer==0.10.0
scikit-learn==1.4.2
scipy==1.13.0
sendgrid==6.11.0
sentry-sdk==1.40.4
six==1.16.0
stack-data==0.6.3
starkbank-ecdsa==2.2.0
stripe==7.10.0
threadpoolctl==3.5.0
tomlkit==0.12.3
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.0.7
wcwidth==0.2.13
Werkzeug==3.0.0
win32-setctime==1.1.0
